Исправлена обработка ошибок

+++ controllers/card.js
@@ -2,6 +2,7 @@ const Card = require('../models/card');
 const NotFoundError = require('../errors/not-found-err');
 const ValidationError = require('../errors/validation-error');
 const NotEnoughRights = require('../errors/not-enough-rights');
+const { validationError, cardNotFound, noRightsToRem, castErr } = require('../errors/error-messages');
 
 // Получить список всех карточек
 
@@ -21,7 +22,8 @@ const createCard = (req, res, next) => {
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        next(new ValidationError('Ошибка валидации'));
+        next(new ValidationError(validationError));
+        return;
       }
       next(err);
     });
@@ -34,8 +36,11 @@ const deleteCard = (req, res, next) => {
   Card.findById(cardId)
     // eslint-disable-next-line consistent-return
     .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
       if (card.owner.toString() !== req.user._id) {
-        throw new NotEnoughRights('Недостаточно прав, чтобы удалить карточку');
+        throw new NotEnoughRights(noRightsToRem);
       }
       Card.findByIdAndRemove(cardId)
         .then(() => res.send(card))
@@ -43,7 +48,8 @@ const deleteCard = (req, res, next) => {
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotFoundError('Карточка не существует'));
+        next(new ValidationError(castErr));
+        return;
       }
       next(err);
     });
@@ -53,10 +59,16 @@ const deleteCard = (req, res, next) => {
 
 const setLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send(card))
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
+      res.send(card);
+    })
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotFoundError('Карточка не существует'));
+        next(new ValidationError(castErr));
+        return;
       }
       next(err);
     });
@@ -66,10 +78,16 @@ const setLike = (req, res, next) => {
 
 const removeLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .then((card) => res.send(card))
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError(cardNotFound);
+      }
+      res.send(card);
+    })
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotFoundError('Карточка не существует'));
+        next(new ValidationError(castErr));
+        return;
       }
       next(err);
     });
